Fix mismatched size of ODP_PACKET_OFFSET_INVALID
ODP_PACKET_OFFSET_INVALID is defined as ((size_t)-1) but the existing
implementations both store offsets using a uint32_t, which breaks things
in a few places on 64-bit systems. There's some explicit casting in the
implementations but it's not consistent and the odp_packet_l2/3/4
functions can return a junk ptr if packet parsing didn't get far enough
to set the offset correctly.

+++ include/odp_packet.h
@@ -30,7 +30,7 @@ typedef uint32_t odp_packet_t;
 #define ODP_PACKET_INVALID ODP_BUFFER_INVALID
 
 /** Invalid offset */
-#define ODP_PACKET_OFFSET_INVALID ((size_t)-1)
+#define ODP_PACKET_OFFSET_INVALID ((uint32_t)-1)
 
 
 /**
